TASK 3: BUSINESS CARDS
This design challenge will allow you to practice the skills mentioned above. Every business relies on business cards as a means of introduction. They are also an expression of the company’s branding. Therefore, it is vital to capture the company’s essence through excellent typography and some ingenuity.

#completed✅💯 

#graphicdesign #graphic #internship #internship2024 #poster #oasisinfobyte
#task_3

complete 💯💯✅