TASK 5: MAGAZINES AND BROCHURES
TASK 5: MAGAZINES AND BROCHURES

Designing the look and feel of a magazine or brochure is a complex project. It involves a range of elements, including advertising, visual representation, photography, art, and graphic design. Working on a magazine design will expose you to the design principles and standards used in print media.


#graphicdesign  #internship #internship2024  #internship #magazines
#oasisinfobyte #task_5 #oasis #intern #graphics #Business